back-cloud/servicio2/requirements.txt
fastapi
uvicorn
pyodbc
sqlalchemy
python-dotenv
bcrypt
pydantic[email]